Добавлена поддрежка нового формата

+++ excelport/src/xlnt/i_format_parser.h
@@ -20,7 +20,7 @@ struct ExcelItem
 	LineType type = LT_NULL;
 	QString gName;
 	QString pName;
-	double rateCount = 0;
+	int rateCount = 0;
 };
 
 class IFormatParser
